Relocate backend-independent helpers out of the IO index package
The chunk planner and two path classifiers are backend-independent but
lived under io.index / io.indexer, forcing core spool behavior to
depend on the IO index package.

- io/index/plan.py -> utils/chunk_plan.py (the planner is used by every
  DataFrameSpool and only imports generic utils).
- is_memory_uri and _directory_writable -> utils/paths.py beside the
  other path classifiers; _directory_writable becomes public
  directory_writable.
- is_memory_uri now matches the exact memorypatch:// / memory://
  schemes instead of any string starting with 'memory', so a real file
  named e.g. memory_notes.h5 is not misclassified.

Imports and doc-reference paths updated; behavior is unchanged.
